refactor(vla): split the repository module
